+++ src/test/php/response/DefaultImageResponseTestCase.php
@@ -84,7 +84,7 @@ class DefaultImageResponseTestCase extends \PHPUnit_Framework_TestCase
     {
         $this->defaultImageResponse->setImage($this->image)->send();
         $this->assertSame($this->handle,
-                          ImageType::$DUMMY->value()->getLastDisplayedHandle()
+                          ImageType::$DUMMY->value()->lastDisplayedHandle()
         );
     }
 
@@ -95,7 +95,7 @@ class DefaultImageResponseTestCase extends \PHPUnit_Framework_TestCase
     public function clearRemovesImageFromResponse()
     {
         $this->defaultImageResponse->setImage($this->image)->clear()->send();
-        $this->assertNull(ImageType::$DUMMY->value()->getLastDisplayedHandle());
+        $this->assertNull(ImageType::$DUMMY->value()->lastDisplayedHandle());
     }
 
     /**